toman/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.2.0
